정자살(classic)·square 엔진 "세로 비율" 슬라이더 소수점 입력 버림 버그 수정
숫자 입력창(numRatio)에서 값을 읽을 때 parseInt를 써서 1.3·1.7 같은
소수점 입력이 정수로 잘렸음(예: 1.3→1). 같은 정수로 잘리는 범위 안에서는
격자 세로 비율이 전혀 안 바뀌어 "고정된 것처럼" 보였던 원인.
parseFloat로 교체. classic 엔진은 JS 쪽 슬라이더 max(3.0)가 실제
HTML/서버 max(5.0)와도 안 맞았어서 5.0으로 맞춤.
(range 슬라이더를 직접 드래그하는 경로는 원래도 정상 동작했음 — 숫자
입력창에 직접 타이핑하는 경로만 문제였음)

+++ src/js/square.js
@@ -1947,7 +1947,7 @@ document.getElementById('muntolColorInput')?.addEventListener('input', e => { se
             draw();
         });
         num.addEventListener('input', () => {
-            const v = parseInt(num.value);
+            const v = parseFloat(num.value);
             if (isNaN(v)) return;
             if (v >= min && v <= max) {
                 range.value = v;
@@ -1955,8 +1955,8 @@ document.getElementById('muntolColorInput')?.addEventListener('input', e => { se
             }
         });
         num.addEventListener('blur', () => {
-            let v = parseInt(num.value);
-            if (isNaN(v)) v = parseInt(range.value);
+            let v = parseFloat(num.value);
+            if (isNaN(v)) v = parseFloat(range.value);
             v = Math.min(max, Math.max(min, v));
             range.value = v;
             num.value   = v;
